+++ lib/cogs/lastfm.py
@@ -90,7 +90,7 @@ class LastFM(Cog):
             return
 
         recent_tracks_url = f"https://ws.audioscrobbler.com/2.0/?method=user.getrecenttracks&user={username}&api_key={os.environ.get('lastfmkey')}&format=json&limit=5"
-    
+        
         async with request("GET", recent_tracks_url) as response:
             data = (await response.json())['recenttracks']
 
@@ -105,9 +105,50 @@ class LastFM(Cog):
 
                 await ctx.send(embed=embed)
 
-
             else:
                 embed=Embed(title=f"⚠ - No recent tracks found for {data['@attr']['user']}.", colour=ctx.author.colour)
+
+                message = await ctx.send(embed=embed)
+            
+                await message.add_reaction("⚠️")
+    
+    @lastfm.group(name="top")
+    async def top_group(self, ctx):
+        if ctx.invoked_subcommand is not "albums":
+            prefix = db.record("SELECT Prefix from guilds WHERE GuildID = ?", ctx.guild.id)
+            await ctx.send(f"Try these commands instead.\n`{prefix[0]}fm top albums`")
+
+    @top_group.command(name="albums")
+    async def top_albums_command(self, ctx, username: Optional[str]):
+        username = username or db.record("SELECT LastfmUsername FROM exp WHERE UserID = ?", ctx.author.id)[0]
+
+        top_albums_url = f"https://ws.audioscrobbler.com/2.0/?method=user.gettopalbums&user={username}&api_key={os.environ.get('lastfmkey')}&format=json&limit=10"
+        User_URL = f"https://ws.audioscrobbler.com/2.0/?method=user.getinfo&user={username}&api_key={os.environ.get('lastfmkey')}&format=json"
+
+        async with request("GET", top_albums_url) as response:
+            data = (await response.json())['topalbums']
+
+            llist = list()
+
+            for i in data['album']:
+                llist.append(f"‣ #{i['@attr']['rank']} - {i['artist']['name']} - {i['name']} ({i['playcount']} plays)")
+
+            if llist:
+                embed=Embed(title=f"{data['@attr']['user']}'s 10 top albums", 
+                            description='\n'.join(llist), colour=ctx.author.colour)
+
+                embed.set_thumbnail(url=data['album'][0]['image'][3]['#text'])
+
+                async with request("GET", User_URL) as icon:
+                    icon_url = (await icon.json())['user']
+
+                    if icon_url['image'][3]['#text'] != "": 
+                        embed.set_footer(text=f"last.fm/users/{data['@attr']['user']}", icon_url=icon_url['image'][3]['#text'])
+
+                await ctx.send(embed=embed)
+
+            else:
+                embed=Embed(title=f"⚠ - No top albums found for {data['@attr']['user']}.", colour=ctx.author.colour)
 
                 message = await ctx.send(embed=embed)
             
@@ -119,9 +160,33 @@ class LastFM(Cog):
 
         await self.fm_search(ctx, username)
 
+    @lastfm.group(name="artist")
+    async def artist_group(self, ctx):
+        if ctx.invoked_subcommand is not "charts":
+            prefix = db.record("SELECT Prefix from guilds WHERE GuildID = ?", ctx.guild.id)
+            await ctx.send(f"Try these commands instead.\n`{prefix[0]}fm artist charts`")
+
+    @artist_group.command(name="charts")
+    async def artist_charts_command(self, ctx):
+        charts_url = f"https://ws.audioscrobbler.com/2.0/?method=chart.gettopartists&api_key={os.environ.get('lastfmkey')}&format=json&limit=10"
+
+        async with request("GET", charts_url) as response:
+            data = (await response.json())['artists']
+
+            llist = list()
+
+            for i in data['artist']:
+                llist.append(f"‣ {i['name']}")
+
+            if llist:
+                embed=Embed(title="Top 10 artists on Last.fm", 
+                            description='\n'.join(llist), colour=ctx.author.colour)
+
+                await ctx.send(embed=embed)
+
     @command(name="setlastfm", aliases=["setfm"], brief="Sets your Last.fm username.")
     @cooldown(1, 5, BucketType.user)
-    async def set_overwatch_profile(self, ctx, username: Optional[str]):
+    async def set_lastfm_username(self, ctx, username: Optional[str]):
         """Sets your Last.fm username for `doob/lastfm`"""
 
         if username != None: